requirements.txt
# K_plus_parcer - Requirements for STAGE 1 (MVP + Markdown Export)
# Updated: 2025-12-18

# ============================================================================
# CORE PDF PROCESSING (PRIORITY!)
# ============================================================================
pymupdf>=1.23.0              # PyMuPDF - базовое извлечение текста из PDF
pymupdf4llm>=0.0.1           # Markdown-friendly экспорт из PDF (ПРИОРИТЕТ!)

# ============================================================================
# DATA VALIDATION & MODELS
# ============================================================================
pydantic>=2.0.0              # Валидация структур данных
python-dateutil>=2.8.0       # Парсинг дат

# ============================================================================
# YAML & JSON
# ============================================================================
pyyaml>=6.0                  # Работа с YAML frontmatter в Markdown

# ============================================================================
# LOGGING & CLI
# ============================================================================
loguru>=0.7.0                # Красивое логирование
click>=8.1.0                 # CLI интерфейс (для будущего расширения)

# ============================================================================
# TESTING (development only)
# ============================================================================
pytest>=7.0.0                # Тестирование
pytest-cov>=4.0.0            # Coverage reports

# ============================================================================
# FUTURE STAGES (commented out for now)
# ============================================================================
# Stage 2: RAG & Table Extraction
# pdfplumber>=0.10.0         # Извлечение таблиц из PDF
# langchain>=0.1.0           # RAG orchestration
# tiktoken>=0.5.0            # Подсчёт токенов

# Stage 3: NLP & Embeddings
# natasha>=1.6.0             # Russian NLP
# pymorphy2>=0.9.0           # Морфологический анализ
# sentence-transformers>=2.0.0  # Embeddings

# Stage 4: ML Datasets
# openai>=1.0.0              # OpenAI API для синтетических данных
# datasets>=2.0.0            # HuggingFace datasets

# Stage 5: Vector Databases
# chromadb>=0.4.0            # ChromaDB
# weaviate-client>=3.0.0     # Weaviate client
